consolidate dcpy test constants and fixtures
adding flexibility to test conenctors with different hard-coded bucket names
